Move server into its own package
